Update pom.xml
Reverted the incorrect GroupID for OrsonPDF so that the repository can be found on Maven Central, but will switch to the newer JFreePDF later today and simply switch it from Java 11 to Java 1.8. It better follows current conventions for Maven use. I did also bump the version number though, as this code base includes amy bugfix that is part of the not-yet-released version 1.9.1.